feat : 비밀번호 기능 추가

+++ toy-project/guestbook/views.py
@@ -10,6 +10,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.http import Http404
+from django.core.exceptions import ValidationError
 
 class PostList(APIView):
     def post(self, request, format=None):
@@ -34,12 +35,20 @@ class PostDetail(APIView):
     def put(self, request, id):
         post = get_object_or_404(Post, id=id)
         serializer = PostSerializer(post, data=request.data)
-        if serializer.is_valid(): #update니까 유효성 검사 필요
-            serializer.save()
-            return Response(serializer.data) 
+        if serializer.is_valid(): 
+            if post.password == request.data.get('password'):
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return JsonResponse({'message':'비밀번호가 올바르지 않습니다.'}, status=400)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request,id):
         post = get_object_or_404(Post, id=id)
-        post.delete()
+        serializer = PostDelSerializer(post, data=request.data)
+        if serializer.is_valid():
+            if post.password == request.data.get('password'):
+                post.delete()
+            else:
+                return JsonResponse({'message':'비밀번호가 올바르지 않습니다.'}, status=400)
         return Response(status=status.HTTP_204_NO_CONTENT)
